+++ components/api-server/src/config.js
@@ -51,7 +51,7 @@ _.merge(config.schema, {
     sessionMaxAge: {
       format: 'duration',
       default: 1000 * 60 * 60 * 24 * 14, // 2 weeks
-      doc: 'The maximum age (in seconds) of a personal access token if unused.'
+      doc: 'The maximum age (in milliseconds) of a personal access token if unused.'
     },
     ssoCookieDomain: {
       format: String,