+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## 1.6
 
+### 1.6.7
+
+Fixes:
+
+- usernames starting with "system" are available
+- personal token expiration now fixed
+
+Removals:
+
+- Deprecated "GET /who-am-i" API method removed
+
 ### 1.6.3
 
 Custom Auth function now has access to all headers.